% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_continuous.R
\name{bin_color_continuous}
\alias{bin_color_continuous}
\title{Bin data and get corresponding colors.}
\usage{
bin_color_continuous(data, by, range, bins, scale = NULL, ascending = TRUE,
  bin_name = "bins")
}
\arguments{
\item{data}{The data to bin. Generally a data table.}

\item{by}{Character; The name of of our output column that we are binning.}

\item{range}{A vector of length two that defines the max and min value of the
scale.}

\item{bins}{Vector; Either a number indicating the number of bins, or a
vector containing the locations where the bins should be cut.}

\item{scale}{List; should be length 4, in the format c(lowest, midpoint
(white), one-bin-above-midpoint, highest).}

\item{ascending}{logical; do increasing values indicate a positive trend?}

\item{bin_name}{Character; Name of column to be created that contains bin
data}
}
\value{
a list \code{x}. \code{x$data} returns the data frame with a new
  column containing the bins, \code{x$colors} contains the corresponding
  color scale.
}
\description{
Bins data for according to a given list.
}
\details{
Currently provides scales and colors for reliability, safeyield, resilience,
and vulnerability. To access, use \code{metric="reliability"},
\code{metric="safeyield"}, etc.
}
\examples{
\dontrun{df <- expand.grid(temp = 0:8, precip = seq(0.7, 1.3, by = 0.1))
df$rel <- seq(40, 100, length=63)
df2 <- bin_color_continuous(df, by = "rel", metric = "reliability")}
}
