% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_response_addin.R
\name{climateResponseAddin}
\alias{climateResponseAddin}
\title{Climate Response Gadget}
\usage{
climateResponseAddin()
}
\value{
A list containing the plot (x$plot), and the data used to create it
  (x$metrics and x$gcm)
}
\description{
Used to easily create climate response surfaces. Can be used on the console,
or via the addins toolbar.
}
