% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CMIP5}
\alias{CMIP5}
\title{CMIP5 data for the Mwache river basin}
\format{An object of class \code{data.frame} with 118 rows and 4 columns.}
\usage{
CMIP5
}
\description{
Provided to show how GCM data can be overlaid on top of a heatmap.
}
\keyword{datasets}
