% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_bins}
\alias{get_bins}
\title{Get bins for metrics.}
\usage{
get_bins(metric)
}
\arguments{
\item{metric}{character; The name of the metric.}
}
\value{
A list containing two sequences. These should be concatenated before
  binning, but are provided separately in order to enable more advanced color
  scales.
}
\description{
Generates a sequence to be used for binning, for a number of different water
resources metrics.
}
\details{
Currently supported metrics: reliability, safeyield.
}

