% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_colors_binary}
\alias{get_colors_binary}
\title{Get colors for binary visualizations.}
\usage{
get_colors_binary()
}
\value{
A list of two colors.
}
\description{
Returns a list of two colors to be used for binary visualizations.
}
\details{
The first value in the list is analagous to "good", "true", etc. The second
is analagous to "bad", "false", etc. Currently this is hardcoded to return
\code{c("royalblue4","firebrick2")}.
}

