% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_colors_continuous}
\alias{get_colors_continuous}
\title{Get colors for continuous visualizations.}
\usage{
get_colors_continuous(baseline = threshold, diverging = TRUE,
  upward = NULL)
}
\arguments{
\item{baseline}{a baseline}

\item{diverging}{logical; does this scale diverge in two directions from the baseline? If scale only increases in one direction, choose TRUE}

\item{upward}{logical; do increasing values indicate a positive trend?}
}
\value{
A list of colors.
}
\description{
Returns a list of colors to be used for continuous visualizations.
}
\details{
Foo
}

