% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_range}
\alias{get_range}
\title{Finds the min and max of the metric.}
\usage{
get_range(data, metric)
}
\value{
vector containing (min_range, max_range)
}
\description{
Finds the min and max of the metric.
}
