% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gcm.R
\name{parse_gcm}
\alias{parse_gcm}
\title{Parse GCM data}
\usage{
parse_gcm(excel_filepath)
}
\arguments{
\item{excel_filepath}{path to excel file}
}
\value{
A dataframe containing four columns (model, scenario, temp, and
  precip), which detail the gcm model, the scenario (e.g. RCP 2.6), the
  temperature change in absolute units, and the preciptation change in
  percentges (where 1 == 0\% change, and 0.9 == -10\% percentage change)
}
\description{
Takes GCM data from Sunwook Wi's GCM analysis tool, and converts it to a data
frame that can be used to overlay GCM information on top of a climate
heatmap.
}
\details{
Sunwook's tool computes monthly averages of temperature and precipitation
change. We average these out to compute a yearly average.
}
