% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gcm.R
\name{parse_sheet_name}
\alias{parse_sheet_name}
\title{Parse GCM sheet names}
\usage{
parse_sheet_name(sheet_name)
}
\arguments{
\item{sheet_name}{string of sheet name}
}
\value{
list(variable, scenario)
}
\description{
Takes potential sheet names from Sunwook Wi's GCM analysis tool, and returns
a list with the variable in question ("temp" or "precip") and the scenario
(e.g. "RCP 2.6").
}
