% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reliability}
\alias{reliability}
\title{Sample dataset}
\format{A dataframe containing 54 observations at various temperatures and
  preciptation and 3 resultant metrics:
  \describe{
     \item{temp}{Temperature change from historical baseline, in degrees celsius}
     \item{precip}{Preciptation change from historical baseline, in percentages}
     \item{reliability}{Reliability, in percentages}
     \item{critical_reliability}{Critical reliability, in percentages}
     \item{safeyield}{Safeyield, in... units}
  }}
\source{
Randomly generated
}
\usage{
reliability
}
\description{
Randomly generated dataset which shows how a metric can be visualized using a
heatmap.
}
\keyword{datasets}
