% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{bin_color_continuous}
\alias{bin_color_continuous}
\title{Bin data and get corresponding colors.}
\usage{
bin_color_continuous(data, by, ascending = T, metric = NULL,
  binName = "bins", numBins = NULL, midpoint = NULL, scale = NULL)
}
\arguments{
\item{data}{The data to bin. Generally a data table.}

\item{by}{Character; The name of the column to bin by.}

\item{ascending}{logical; do increasing values indicate a positive trend?}

\item{metric}{character; The name of the metric we want to create a color
scale for.}

\item{binName}{Character; Name of column to be created that contains bin data}

\item{numBins}{Vector; Provide a vector of length 2, where the first number
is the number of bins below the threshold (inclusive), and the second
number is the number of bins above the treshold.}

\item{midpoint}{Only used if binary == FALSE. Sets the value where the color
scale diverges. If NULL, defaults to middle of metric range.}

\item{scale}{List; should be length 4, in the format c(lowest, midpoint,
one-bin-above-midpoint, highest).}

\item{bins}{A list that defines the bins you would like to create.}
}
\value{
a list \code{x}. \code{x$data} returns the data frame with a new
  column containing the bins, \code{x$colors} contains the corresponding
  color scale.
}
\description{
Bins data for according to a given list.
}
\details{
Currently provides scales and colors for reliability, safeyield, resilience,
and vulnerability. To access, use \code{metric="reliability"},
\code{metric="safeyield"}, etc.
}
\examples{
\dontrun{df <- expand.grid(temp = 0:8, precip = seq(0.7, 1.3, by = 0.1))
df$rel <- seq(40, 100, length=63)
df2 <- bin_color_continuous(df, by = "rel", metric = "reliability")}
}
