% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_continuous.R
\name{climate_heatmap_continuous}
\alias{climate_heatmap_continuous}
\title{Create a climate heatmap.}
\usage{
climate_heatmap_continuous(data, metric, bins = 7, ascending = TRUE,
  range = NULL, colors = NULL, to_percent = c(FALSE, TRUE),
  z_axis_title = "Range")
}
\arguments{
\item{data}{The data. dataframe, data table, named matrix, etc.}

\item{metric}{character; The metric to display in the z axis. There should be
a corresponding column name.}

\item{bins}{Either a single number indicating the number of bins, or a vector
that specifies where the bins will be cut.}

\item{ascending}{logical; Do increasing values in the metric indicate
increasingly acceptable performance?}

\item{range}{A vector of length two that defines the max and min value of the
scale.}

\item{colors}{Provide a list representing a color scale.}

\item{to_percent}{list; List of length two that says whether or not (temp,
precip) data should be treated as percent changes (e.g. 1 means 0% change,
0.9 means -10% change, 1.2 means 20% change...)}
}
\value{
A ggplot2 object representing the heatmap.
}
\description{
Creates a ggplot2 heatmap with temperature on the x axis, preciptation on the
y axis, and a given metric on the z axis.
}
\details{
In order for this function to work, you must have a column named "temp" and
another named "precip". Beyond that, you can have as many columns as you'd
like.

Because this is a ggplot2 object, you should be able to overwrite many of
settings defined here.
}
\examples{
df <- expand.grid(temp=0:8,precip=seq(0.7,1.3,by=0.1))
df$rel <- seq(40,100,length=63)
climate_heatmap_continuous(df,"reliability")

}
